♻️ [refactor]: #452 Phase 9 edge: lexer_read_real_edge.rs を分離
read_real のエッジテスト（Phase 9-G 指数拒否 / 9-H IEEE 754 ±0.0 /
9-I pos 巻き戻し / 9-J 大桁数 / 9-K dispatcher 契約）を
`src/lexer/tests/lexer_read_real_edge.rs` に分離。境界は
旧 lexer.rs:2043-2360 の行範囲（Phase 9-G から Phase 10 ヘッダ手前まで）。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_real_edge.rs: 319 行 (< 800)

Refs #452
